Fixed linter. Rollback NewBatchProcessor method signature

+++ sequencer/sequencer.go
@@ -95,7 +95,7 @@ func (s *Sequencer) Start() {
 				log.Warnf("failed to get last batch from the state, err: %v", err)
 				continue
 			}
-			bp, err := s.State.NewBatchProcessor(int64(lastBatch.BatchNumber), false)
+			bp, err := s.State.NewBatchProcessor(lastBatch.BatchNumber, false)
 			if err != nil {
 				log.Warnf("failed to create new batch processor, err: %v", err)
 				continue
